feat: 적금 계좌 생성 및 인출 기능 구현
[SavingsAccountService]

- 동시성 이슈 고려
    - 메인 계좌 충전과 정기 적금 인출이 동시에 호출되는 상황을 고려해, 비관적 락(쓰기 락) 사용
- 트랜잭션 격리 레벨 고려
    - findByMemberIdWithWriteLock() 호출 시 memberId 에 인덱스 추가
    - 트랜잭션 격리 레벨 2를 사용해서, update 호출 시 where 조건에 부합하는 레코드를 찾기 전까지 마주치는 모든 계좌 레코드에 락을 바로 해제 시켜서 다른 트랜잭션에서 접근할 수 있도록 함
        - 갭락을 해제함으로써 phantom read 현상이 발생할 수 있지만 이로 인한 추가적인 이슈는 없을 것이라 예상
- 자동으로 수행되는 정기 적금 API와, 자유롭게 수행되는 자유 적금 API 분리

+++ src/test/java/org/c4marathon/assignment/service/AccountServiceTest.java
@@ -3,7 +3,6 @@ package org.c4marathon.assignment.service;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import org.c4marathon.assignment.common.exception.BusinessException;
@@ -55,13 +54,13 @@ class AccountServiceTest {
 		var accountId = 1L;
 		var chargeAmount = 1000;
 		var accumulatedChargeAmount = 100000;
-		var totalAmount = 1000;
+		var totalAmount = 1000L;
 		var account = mock(Account.class);
 
 		given(accountRepository.findByIdWithWriteLock(anyLong())).willReturn(Optional.of(account));
 		given(account.getChargeLimit()).willReturn(ChargeLimit.DAY_BASIC_LIMIT);
 		given(account.getAccumulatedChargeAmount()).willReturn(accumulatedChargeAmount);
-		given(account.getAmount()).willReturn(BigInteger.valueOf(totalAmount));
+		given(account.getAmount()).willReturn(totalAmount);
 
 		// when
 		var result = accountService.charge(accountId, chargeAmount);
